+++ NEWS.md
@@ -1,5 +1,8 @@
 # attachment 0.0.3.9000
 
+* `att_to_description(add_version = TRUE)` adds version of package in DESCRIPTION
+* `att_to_description(pkg_ignore)` adds possibility to ignore some packages
+
 # attachment 0.0.3
 
 * Get ready for CRAN